fix: resolve remaining module ambiguity and unused import errors
- Remove stale schedule/tests.rs and db/tests.rs (split dirs exist)
- Add #[cfg(test)] re-exports for test-only types in alerts and sessions
- Apply cargo fmt formatting fixes
