+++ crates/opengoose-web/src/handlers/alerts/mod.rs
@@ -6,6 +6,7 @@ mod test_run;
 
 pub use listing::{alert_history, list_alerts};
 pub use mutations::{create_alert, delete_alert};
+#[cfg(test)]
 pub use requests::{AlertHistoryQueryParams, CreateAlertRequest, TestAlertQueryParams};
 #[allow(unused_imports)]
 pub use responses::{AlertHistoryResponse, AlertRuleResponse};
