+++ crates/opengoose-web/src/handlers/sessions/mod.rs
@@ -3,8 +3,12 @@ mod listing;
 mod messages;
 mod models;
 
-pub use listing::{ListQuery, list_sessions};
-pub use messages::{MessagesQuery, get_messages};
+#[cfg(test)]
+pub use listing::ListQuery;
+pub use listing::list_sessions;
+#[cfg(test)]
+pub use messages::MessagesQuery;
+pub use messages::get_messages;
 pub use models::{MessageItem, SessionItem};
 
 #[cfg(test)]
